# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orders/my-orders": {
      "get": {
        "operationId": "OrdersController_findMyOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderSummaryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user orders",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/my-orders/{orderNumber}": {
      "get": {
        "operationId": "OrdersController_findMyOrderByOrderNumber",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user order by order number",
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/guest/lookup": {
      "post": {
        "operationId": "OrdersController_findGuestOrder",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Guest: lookup order by order number and email",
        "tags": [
          "Orders"
        ]
      }
    },
    "/admin/orders": {
      "get": {
        "operationId": "OrdersController_findAllAdminOrders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderSummaryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: get all orders",
        "tags": [
          "Orders"
        ]
      }
    },
    "/admin/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findAdminOrderById",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: get order by ID",
        "tags": [
          "Orders"
        ]
      }
    },
    "/admin/orders/{id}/status": {
      "patch": {
        "operationId": "OrdersController_updateAdminOrderStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Admin: update order status",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zemlo API",
    "description": "Zemlo e-commerce backend API contract",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "OrderProductSummaryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "OrderVariantSummaryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "OrderItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "productId": {
            "type": "string"
          },
          "variantId": {
            "type": "string",
            "nullable": true
          },
          "productName": {
            "type": "string"
          },
          "variantName": {
            "type": "string",
            "nullable": true
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "number"
          },
          "totalPrice": {
            "type": "number"
          },
          "productSnapshot": {
            "type": "object",
            "nullable": true
          },
          "product": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderProductSummaryResponseDto"
              }
            ]
          },
          "variant": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderVariantSummaryResponseDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderId",
          "productId",
          "productName",
          "quantity",
          "unitPrice",
          "totalPrice",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderPaymentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "enum": [
              "STRIPE",
              "PAYPAL",
              "CREDIT_CARD",
              "DEBIT_CARD",
              "BANK_TRANSFER",
              "CASH_ON_DELIVERY",
              "MANUAL"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "FAILED",
              "REFUNDED",
              "PARTIALLY_REFUNDED"
            ]
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "transactionId": {
            "type": "string",
            "nullable": true
          },
          "paymentIntentId": {
            "type": "string",
            "nullable": true
          },
          "paidAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "failedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "refundedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "failureReason": {
            "type": "string",
            "nullable": true
          },
          "gatewayResponse": {
            "type": "object",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderId",
          "method",
          "status",
          "amount",
          "currency",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderSummaryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "guestEmail": {
            "type": "string",
            "nullable": true
          },
          "guestPhone": {
            "type": "string",
            "nullable": true
          },
          "guestFirstName": {
            "type": "string",
            "nullable": true
          },
          "guestLastName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED",
              "REFUNDED"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "FAILED",
              "REFUNDED",
              "PARTIALLY_REFUNDED"
            ]
          },
          "fulfillmentStatus": {
            "type": "string",
            "enum": [
              "UNFULFILLED",
              "PARTIALLY_FULFILLED",
              "FULFILLED"
            ]
          },
          "subtotal": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "shippingCost": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "customerNote": {
            "type": "string",
            "nullable": true
          },
          "shippingMethod": {
            "type": "string",
            "nullable": true
          },
          "shippingCarrier": {
            "type": "string",
            "nullable": true
          },
          "trackingNumber": {
            "type": "string",
            "nullable": true
          },
          "trackingUrl": {
            "type": "string",
            "nullable": true
          },
          "paidAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "estimatedDelivery": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "actualDelivery": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "payment": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderPaymentResponseDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "paymentStatus",
          "fulfillmentStatus",
          "subtotal",
          "tax",
          "shippingCost",
          "discount",
          "total",
          "items",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderCustomerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "STAFF",
              "ADMIN",
              "SUPER_ADMIN"
            ]
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "role"
        ]
      },
      "OrderAddressResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "SHIPPING",
              "BILLING",
              "BOTH"
            ]
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "apartment": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderId",
          "type",
          "firstName",
          "lastName",
          "phone",
          "street",
          "city",
          "state",
          "zipCode",
          "country",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderStatusHistoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED",
              "REFUNDED"
            ]
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "changedBy": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderId",
          "status",
          "createdAt"
        ]
      },
      "OrderDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "guestEmail": {
            "type": "string",
            "nullable": true
          },
          "guestPhone": {
            "type": "string",
            "nullable": true
          },
          "guestFirstName": {
            "type": "string",
            "nullable": true
          },
          "guestLastName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "PROCESSING",
              "SHIPPED",
              "DELIVERED",
              "CANCELLED",
              "REFUNDED"
            ]
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "FAILED",
              "REFUNDED",
              "PARTIALLY_REFUNDED"
            ]
          },
          "fulfillmentStatus": {
            "type": "string",
            "enum": [
              "UNFULFILLED",
              "PARTIALLY_FULFILLED",
              "FULFILLED"
            ]
          },
          "subtotal": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "shippingCost": {
            "type": "number"
          },
          "discount": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "customerNote": {
            "type": "string",
            "nullable": true
          },
          "shippingMethod": {
            "type": "string",
            "nullable": true
          },
          "shippingCarrier": {
            "type": "string",
            "nullable": true
          },
          "trackingNumber": {
            "type": "string",
            "nullable": true
          },
          "trackingUrl": {
            "type": "string",
            "nullable": true
          },
          "paidAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "completedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "estimatedDelivery": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "actualDelivery": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponseDto"
            }
          },
          "payment": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderPaymentResponseDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderCustomerResponseDto"
              }
            ]
          },
          "shippingAddress": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderAddressResponseDto"
              }
            ]
          },
          "billingAddress": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/OrderAddressResponseDto"
              }
            ]
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderStatusHistoryResponseDto"
            }
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "paymentStatus",
          "fulfillmentStatus",
          "subtotal",
          "tax",
          "shippingCost",
          "discount",
          "total",
          "items",
          "createdAt",
          "updatedAt",
          "statusHistory"
        ]
      },
...
    }
  }
}